+++ src/main/java/com/investment/metal/database/ExpressionFunctionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExpressionFunctionRepository extends JpaRepository<ExpressionFunction, Long> {
+public interface ExpressionFunctionRepository extends JpaRepository<ExpressionFunction, Integer> {
 
 }
