+++ src/main/java/com/investment/metal/database/MetalPriceRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface MetalPriceRepository extends JpaRepository<MetalPrice, Long> {
+public interface MetalPriceRepository extends JpaRepository<MetalPrice, Integer> {
 
     @Query("select t from MetalPrice t where t.metalSymbol=?1 order by t.id desc")
     Optional<List<MetalPrice>> findByMetalSymbol(String symbol);
