+++ src/main/java/com/investment/metal/database/Purchase.java
@@ -13,9 +13,9 @@ public class Purchase {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
-    private Long userId;
+    private Integer userId;
 
     private String metalSymbol;
