+++ src/main/java/com/investment/metal/database/RevolutProfitRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RevolutProfitRepository extends JpaRepository<RevolutProfit, Long> {
+public interface RevolutProfitRepository extends JpaRepository<RevolutProfit, Integer> {
     @Query(value="SELECT * FROM RevolutProfit WHERE metal_symbol=?1 ORDER BY time DESC LIMIT 1", nativeQuery = true)
     RevolutProfit findByMetalSymbol(String metalSymbol);
